Move main_page to index
